fix last commit, i think?

+++ server/controls/es.py
@@ -14,8 +14,8 @@ for _e in events:
         controller.press(key, type=numType, shift=shift)
     else: #mouse event
         #move, set
-        if e[1] == "mm": controller.move_mouse(int(e[2]), int(e[3]))
-        if e[1] == "sm": controller.set_mouse(int(e[2]), int(e[3]))
+        if e[1] == "mm": controller.move_mouse(float(e[2]), float(e[3]))
+        if e[1] == "sm": controller.set_mouse(float(e[2]), float(e[3]))
         #press
         if e[1].startswith('p'): controller.press_mouse(True if e[1][1] == "r" else False)
         #release
